Refactored the structure of our files to make things alot easier to navigate around
