+++ app/javascript/packs/application.js
@@ -16,6 +16,6 @@ require("channels")
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
 // Support component names relative to this directory:
-var componentRequireContext = require.context("components", true);
+var componentRequireContext = require.context("react-components", true);
 var ReactRailsUJS = require("react_ujs");
 ReactRailsUJS.useContext(componentRequireContext);
