+++ app/javascript/react-components/components/Header.js
@@ -13,7 +13,6 @@ function Header(props) {
     return (
         <React.Fragment>
             <nav>
-            {logged_in &&
                 <nav className="navbar navbar-expand-lg navbar-dark bg-dark">
                   <a className="navbar-brand" href="/"><b>KALI</b></a>
                   <button className="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarColor02" aria-controls="navbarColor02" aria-expanded="false" aria-label="Toggle navigation">
@@ -33,7 +32,6 @@ function Header(props) {
                       </li>
 
                     </ul>
-                }
                     <form className="form-inline my-2 my-lg-0">
                       {logged_in &&
                         <div>
