+++ app/javascript/react-components/components/about-us/ianthompson.jpg
