+++ app/javascript/react-components/components/home-carousel/HomeCarousel.js
