+++ app/javascript/react-components/pages/career-main/CareerMainPage.js
@@ -1,7 +1,7 @@
 import React from "react"
 import PropTypes from "prop-types"
 import { Link } from 'react-router-dom'
-import MainTaskList from './pages/tasks/tasklist/MainTaskList'
+import MainTaskList from '../../components/MainTaskList'
 import "bootswatch/dist/lux/bootstrap.min.css";
 import 'bootstrap/dist/css/bootstrap.min.css';
