+++ app/javascript/react-components/pages/home/HomePage.js
@@ -1,7 +1,7 @@
 import React from "react"
 import PropTypes from "prop-types"
-import HomeCarousel from './HomeCarousel.js'
-import AboutUs from './AboutUs.js'
+import HomeCarousel from '../../components/home-carousel/HomeCarousel.js'
+import AboutUs from '../../components/about-us/AboutUs.js'
 
 
 function HomePage() {
