+++ app/javascript/react-components/pages/jobs/CurrentJobCard.js
