+++ app/javascript/react-components/pages/jobs/ShowCurrentJob.js
